+++ client/src/main/java/ar/edu/itba/pod/client/query5/Query5Client.java
@@ -24,7 +24,7 @@ public class Query5Client extends QueryClient {
 
     private final IMap<String, Infraction> infractionsMap;
 
-    private final MultiMap<String, Integer> ticketsMap;
+    private final MultiMap<String, Double> ticketsMap;
 
     private final IMap<String, Integer> auxMap;
     public Query5Client(String query) {
@@ -51,8 +51,8 @@ public class Query5Client extends QueryClient {
 
     public SortedSet<Query5Result> executeJob() throws ExecutionException, InterruptedException {
         final JobTracker tracker = this.hazelcast.getJobTracker(Util.HAZELCAST_NAMESPACE);
-        final KeyValueSource<String,Integer> source = KeyValueSource.fromMultiMap(ticketsMap);
-        final Job<String, Integer> job = tracker.newJob(source);
+        final KeyValueSource<String,Double> source = KeyValueSource.fromMultiMap(ticketsMap);
+        final Job<String, Double> job = tracker.newJob(source);
 
         Map<String, Integer> aux = job
                 .mapper(new Query5FirstMapper())
